requirements.txt
# ══════════════════════════════════════════════════════════════════════════════
# INDUSTRY PRACTICUM - PARTNERLINQ
# Python Dependencies
# ══════════════════════════════════════════════════════════════════════════════
# Install with: pip install -r requirements.txt
# Or in development mode: pip install -r requirements.txt --upgrade
# ══════════════════════════════════════════════════════════════════════════════

# ── Core AI & LLM ─────────────────────────────────────────────────────────────
# Groq SDK (still used as fallback for environments without the openai package)
groq~=1.0.0

# OpenAI SDK — also used for Groq (OpenAI-compatible) and NVIDIA NIM endpoints
openai>=1.0.0

# Anthropic SDK — required only when provider="anthropic"
anthropic>=0.25.0

# Environment variable management (loads .env file for API keys)
python-dotenv~=1.2.0

# ── EDI & Data Format Parsing ─────────────────────────────────────────────────
# EDIFACT (Electronic Data Interchange For Administration, Commerce and Transport)
pydifact~=0.1.8

# XML/XSLT parsing and processing (lxml with XPath and XSLT support)
lxml~=6.0.0

# Saxon/C Python wrapper — XSLT 2.0 and 3.0 execution, no JVM required
# Used as the primary XSLT engine in simulate; lxml is the 1.0 fallback
saxonche>=12.0.0

# ── RAG / Multi-File Engine ────────────────────────────────────────────────────
# Local persistent vector store — no server required
chromadb>=0.6.0

# Local sentence embedding model (all-MiniLM-L6-v2, ~80 MB) — no API key needed
sentence-transformers>=3.0.0

# ── Frontend ───────────────────────────────────────────────────────────────────
# Streamlit web UI — file upload, chat interface, audit form, RAG tab
streamlit>=1.32.0

# ══════════════════════════════════════════════════════════════════════════════
# Version Constraints Explanation:
# ~= (compatible release): Allows patch-level changes (e.g., ~=1.0.0 allows 1.0.x)
# >= (minimum version): Use when you need at least a specific version
# == (exact version): Use only when absolutely necessary for compatibility
# ══════════════════════════════════════════════════════════════════════════════